requirements.txt
jupyter
requests
pillow
beautifulsoup4
qrcode
algorithms
